remove extra methods

+++ banking-system-main/src/main/java/com/javahongkong/bootcamp/exercise/Account.java
@@ -29,7 +29,7 @@ public abstract class Account implements AccountInterface {
 		return balance;
 	}
 
-	public int getPin() {
+	public double getPin() {
 		return pin;
 	}
